+++ src/core/Header/index.ts
@@ -1,0 +1,16 @@
+/*---------------------------------------------------------------------------------------------
+ * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+ * See LICENSE.md in the project root for license terms and full copyright notice.
+ *--------------------------------------------------------------------------------------------*/
+export { Header } from './Header';
+export type { HeaderProps } from './Header';
+export default './Header';
+
+export { HeaderButton } from './HeaderButton';
+export type { HeaderButtonProps } from './HeaderButton';
+
+export { HeaderLogo } from './HeaderLogo';
+export type { HeaderLogoProps } from './HeaderLogo';
+
+export { HeaderBreadcrumbs } from './HeaderBreadcrumbs';
+export type { HeaderBreadcrumbsProps } from './HeaderBreadcrumbs';